+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 - [#548](https://github.com/mobilityhouse/ocpp/issues/548) OCPP 2.0.1 MessageInfoType attribute name correction
 - [#300](https://github.com/mobilityhouse/ocpp/issues/300) OCPP 2.0.1 add reference components and variables
 - [#518](https://github.com/mobilityhouse/ocpp/issues/518) OCPP 2.0.1 add additional reason codes from v1.3
+- [#569](https://github.com/mobilityhouse/ocpp/issues/569) Revert DataTransfer.data type hint
 
 ## 0.24.0 (2023-12-07)